+++ lib/src/address_ipv4.dart
@@ -1,8 +1,6 @@
-/// [IPv4Address] is a helper class to ensure that your IPv4 address has been formatted correctly.
+/// [IPv4Address] helps ensure that your IPv4 address has been formatted correctly.
 ///
 /// The class has a static function, `validate(String address)` which allows easy validation that an IPv4 address string is correctly formatted.
-///
-/// Create an [IPv4Address] object by using the factory `IPv4Address.from(address)` where address is a string representation of the address. The factory will call the validation function mentioned above, but will throw a [FormatException] on a poorly constructed string, so it is recommended to validate it first.
 class IPv4Address {
   static const String _regex =
       r"\b((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)(\.|$)){4}\b";
@@ -18,7 +16,7 @@ class IPv4Address {
   /// When using for Wake on LAN functionality, please ensure this address is the broadcast IPv4 address of your network.
   /// This typically is the IPv4 address of your machine with the last block set to 255.
   /// _This can differ depending on the subnet mask used for the network_.
-  factory IPv4Address.from(String address) {
+  factory IPv4Address(String address) {
     if (!IPv4Address.validate(address)) {
       throw FormatException('Not a valid IPv4 address string');
     }