+++ test/address_ipv4_test.dart
@@ -4,13 +4,13 @@ import 'package:test/test.dart';
 void main() {
   _getters();
   _functionValidate();
-  _factoryFrom();
+  _constructor();
 }
 
 void _getters() {
   group('Getters', () {
     String address = '192.168.1.1';
-    IPv4Address ipv4 = IPv4Address.from(address);
+    IPv4Address ipv4 = IPv4Address(address);
     test('address', () {
       expect(ipv4.address, address);
     });
@@ -42,23 +42,23 @@ void _functionValidate() {
   });
 }
 
-void _factoryFrom() {
-  group('Factory: .from()', () {
+void _constructor() {
+  group('Constructor', () {
     test('Valid Instance', () {
       String address = '192.168.1.1';
-      expect(IPv4Address.from(address), equals(isA<IPv4Address>()));
+      expect(IPv4Address(address), equals(isA<IPv4Address>()));
     });
     test('Invalid Instance: Empty String', () {
       String address = '';
-      expect(() => IPv4Address.from(address), throwsA(isA<FormatException>()));
+      expect(() => IPv4Address(address), throwsA(isA<FormatException>()));
     });
     test('Invalid Instance: Invalid String (Alpha)', () {
       String address = 'hello';
-      expect(() => IPv4Address.from(address), throwsA(isA<FormatException>()));
+      expect(() => IPv4Address(address), throwsA(isA<FormatException>()));
     });
     test('Invalid Instance: Invalid String (MAC structure, invalid octet)', () {
       String address = '192.168.1.256';
-      expect(() => IPv4Address.from(address), throwsA(isA<FormatException>()));
+      expect(() => IPv4Address(address), throwsA(isA<FormatException>()));
     });
   });
 }