+++ test/address_mac_test.dart
@@ -4,14 +4,14 @@ import 'package:test/test.dart';
 void main() {
   _getters();
   _functionValidate();
-  _factoryFrom();
+  _constructor();
 }
 
 void _getters() {
   group('Getters', () {
     String address = '00:FF:00:FF:00:FF';
     List<int> bytes = [0, 255, 0, 255, 0, 255];
-    MACAddress mac = MACAddress.from(address);
+    MACAddress mac = MACAddress(address);
     test('address', () {
       expect(mac.address, address);
     });
@@ -50,27 +50,27 @@ void _functionValidate() {
   });
 }
 
-void _factoryFrom() {
-  group('Factory: .from()', () {
+void _constructor() {
+  group('Constructor', () {
     test('Valid Instance', () {
       String address = '00:00:00:00:00:00';
-      expect(MACAddress.from(address), equals(isA<MACAddress>()));
+      expect(MACAddress(address), equals(isA<MACAddress>()));
     });
     test('Invalid Instance: Empty String', () {
       String address = '';
-      expect(() => MACAddress.from(address), throwsA(isA<FormatException>()));
+      expect(() => MACAddress(address), throwsA(isA<FormatException>()));
     });
     test('Invalid Instance: Invalid String (Alpha)', () {
       String address = 'hello';
-      expect(() => MACAddress.from(address), throwsA(isA<FormatException>()));
+      expect(() => MACAddress(address), throwsA(isA<FormatException>()));
     });
     test('Invalid Instance: Invalid String (MAC structure, invalid octet)', () {
       String address = '00:00:00:00:FG:00';
-      expect(() => MACAddress.from(address), throwsA(isA<FormatException>()));
+      expect(() => MACAddress(address), throwsA(isA<FormatException>()));
     });
     test('Invalid Instance: Invalid String (MAC structure, invalid octet)', () {
       String address = '00:00:00:00:A:00';
-      expect(() => MACAddress.from(address), throwsA(isA<FormatException>()));
+      expect(() => MACAddress(address), throwsA(isA<FormatException>()));
     });
   });
 }